Inline missing support types for build
